rename spectest.go
