+++ shared/testutil/BUILD.bazel
@@ -6,10 +6,10 @@ go_library(
     testonly = True,
     srcs = [
         "attestation.go",
+        "bazel.go",
         "block.go",
         "deposits.go",
         "helpers.go",
-        "spectest.go",
         "state.go",
         "wait_timeout.go",
     ],
@@ -27,20 +27,14 @@ go_library(
         "//shared/interop:go_default_library",
         "//shared/params:go_default_library",
         "//shared/rand:go_default_library",
-        "//shared/testutil/require:go_default_library",
         "//shared/trieutil:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
-        "@com_github_ghodss_yaml//:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
-        "@com_github_golang_snappy//:go_default_library",
-        "@com_github_json_iterator_go//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@in_gopkg_d4l3k_messagediff_v1//:go_default_library",
         "@io_bazel_rules_go//go/tools/bazel:go_default_library",
     ],
 )
